requirements.txt
fastapi==0.100.0
uvicorn==0.25.0
google-auth==2.23.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.100.0
pyjwt==2.7.0
psycopg2==2.9.7
python-dotenv==1.0.0

# For testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.0


# Agents
openai==0.27.2